+++ src/primitives/PieceImage/PieceDecorations/index.ts
@@ -1,1 +1,2 @@
-export * from "./PieceDecorations";
+export * from "./DecorationsAbovePiece";
+export * from "./DecorationsBelowPiece";
